+++ src/app/cotizar/accidentes-personales/page.tsx
@@ -1,0 +1,5 @@
+import FormAccidentesPersonales from '@/components/cotizar/seguros/form-accidentes-personales'
+
+export default function Page() {
+  return <FormAccidentesPersonales />
+}