+++ src/app/cotizar/art/page.tsx
@@ -1,0 +1,5 @@
+import FormArt from '@/components/cotizar/seguros/form-art'
+
+export default function Page() {
+  return <FormArt />
+}